requirements.txt
requests>=2.31
pandas>=2.0
beautifulsoup4>=4.12
python-dotenv>=1.0
lxml>=4.9
